moved sortNotes over to note utils
